Support looping constructs, plus a refactor.
Specifically for environment handling. It turns out that the way I
previously was dealing with environments (using a Treemap and creating a
plumbing through copies) does not really work well with the loop
constructs (by default). Changing the environment to a linked list of
hashtables solved this probelm, and also cleaned up the code somewhat.