separate global environment and fix return
Apparently, the semantics of Lox specify that the global environment
should be treated differently from local ones. In particular, we allow
closures to see new bindings in the global environment, but not in the
local one. With this, we can perform mutual recursion, as long as it's at
the top level.

I accidentally joined these two commits and can't be bothered to fix it.
The second change makes it so that returning when not in a function
results in a run-time error. This behavior used to exist before the
monadic refactor, though with a less jank implementation.